+++ app/modules/admin/controllers/Controller.php
@@ -6,8 +6,6 @@ use yii\filters\AccessControl;
 
 class Controller extends \yii\web\Controller
 {
-	public $layout = 'main';
-
 	/**
 	 * @inheritdoc
 	 */