+++ app/modules/admin/views/partials/footer.php
@@ -1,0 +1,4 @@
+
+<footer class="main-footer">
+	<strong>&copy; 2017 My Company</strong>. All rights reserved.
+</footer>